requirements.txt
streamlit
openai
langchain
pandas
matplotlib
seaborn
plotly
python-dotenv
